requirements.txt
Pillow==10.4.0
ultralytics==8.2.89
